+++ jax/scipy/special.py
@@ -21,8 +21,9 @@ from .. import util
 from .. import lax
 from .. import api
 from ..numpy import lax_numpy as jnp
-from ..numpy.lax_numpy import (_wraps, asarray, _reduction_dims, _constant_like,
+from ..numpy.lax_numpy import (asarray, _reduction_dims, _constant_like,
                                _promote_args_inexact)
+from ..numpy._util import _wraps                        
 
 
 @_wraps(osp_special.gammaln)
