+++ jax/scipy/stats/bernoulli.py
@@ -17,10 +17,11 @@ import scipy.stats as osp_stats
 
 from ... import lax
 from ...numpy import lax_numpy as jnp
+from ...numpy._util import _wraps
 from ..special import xlogy, xlog1py
 
 
-@jnp._wraps(osp_stats.bernoulli.logpmf, update_doc=False)
+@_wraps(osp_stats.bernoulli.logpmf, update_doc=False)
 def logpmf(k, p, loc=0):
   k, p, loc = jnp._promote_args_inexact("bernoulli.logpmf", k, p, loc)
   zero = jnp._constant_like(k, 0)
@@ -30,6 +31,6 @@ def logpmf(k, p, loc=0):
   return jnp.where(jnp.logical_or(lax.lt(x, zero), lax.gt(x, one)),
                   -jnp.inf, log_probs)
 
-@jnp._wraps(osp_stats.bernoulli.pmf, update_doc=False)
+@_wraps(osp_stats.bernoulli.pmf, update_doc=False)
 def pmf(k, p, loc=0):
   return jnp.exp(logpmf(k, p, loc))
