+++ jax/scipy/stats/gamma.py
@@ -15,7 +15,8 @@
 import scipy.stats as osp_stats
 
 from ... import lax
-from ...numpy.lax_numpy import (_promote_args_inexact, _constant_like, _wraps,
+from ...numpy._util import _wraps
+from ...numpy.lax_numpy import (_promote_args_inexact, _constant_like,
                                 where, inf)
 from ..special import gammaln
