+++ jax/scipy/stats/pareto.py
@@ -16,7 +16,8 @@
 import scipy.stats as osp_stats
 
 from ... import lax
-from ...numpy.lax_numpy import _promote_args_inexact, _constant_like, _wraps, inf, where
+from ...numpy._util import _wraps
+from ...numpy.lax_numpy import _promote_args_inexact, _constant_like, inf, where
 
 
 @_wraps(osp_stats.pareto.logpdf, update_doc=False)
