+++ jax/scipy/stats/poisson.py
@@ -16,11 +16,12 @@
 import scipy.stats as osp_stats
 
 from ... import lax
+from ...numpy._util import _wraps
 from ...numpy import lax_numpy as jnp
 from ..special import xlogy, gammaln
 
 
-@jnp._wraps(osp_stats.poisson.logpmf, update_doc=False)
+@_wraps(osp_stats.poisson.logpmf, update_doc=False)
 def logpmf(k, mu, loc=0):
   k, mu, loc = jnp._promote_args_inexact("poisson.logpmf", k, mu, loc)
   zero = jnp._constant_like(k, 0)
@@ -28,6 +29,6 @@ def logpmf(k, mu, loc=0):
   log_probs = xlogy(x, mu) - gammaln(x + 1) - mu
   return jnp.where(lax.lt(x, zero), -jnp.inf, log_probs)
 
-@jnp._wraps(osp_stats.poisson.pmf, update_doc=False)
+@_wraps(osp_stats.poisson.pmf, update_doc=False)
 def pmf(k, mu, loc=0):
   return jnp.exp(logpmf(k, mu, loc))
