+++ jax/scipy/stats/t.py
@@ -17,7 +17,8 @@ import numpy as np
 import scipy.stats as osp_stats
 
 from ... import lax
-from ...numpy.lax_numpy import _promote_args_inexact, _constant_like, _wraps
+from ...numpy._util import _wraps
+from ...numpy.lax_numpy import _promote_args_inexact, _constant_like
 
 
 @_wraps(osp_stats.t.logpdf, update_doc=False)
