+++ jax/scipy/stats/uniform.py
@@ -16,7 +16,8 @@
 import scipy.stats as osp_stats
 
 from ... import lax
-from ...numpy.lax_numpy import (_constant_like, _promote_args_inexact, _wraps,
+from ...numpy._util import _wraps
+from ...numpy.lax_numpy import (_constant_like, _promote_args_inexact,
                                 where, inf, logical_or)
